+++ jina/peapods/__init__.py
@@ -1,6 +1,4 @@
-
-
 from .peas import BasePea as Pea
 from .pods import BasePod
 from .pods import Pod
-from .pods.compoundpod import CompoundPod
+from .pods.compound import CompoundPod
